% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\name{AllPrimesUpTo}
\alias{AllPrimesUpTo}
\title{Get all prime numbers.}
\usage{
AllPrimesUpTo(n)
}
\arguments{
\item{n}{A positive integer.}
}
\value{
A integer vector of all prime numbers up to n.
}
\description{
Get all prime numbers up to n.
}
\examples{
AllPrimesUpTo(10)
AllPrimesUpTo(100)
AllPrimesUpTo(1e6)
}
