% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primefactr.R
\name{primefactr-package}
\alias{primefactr-package}
\alias{primefactr}
\title{R package that uses Prime Factorization for computations.}
\arguments{
\item{n}{A positive integer.}
}
\description{
TODO
}
\keyword{package}
