% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\name{IsPrime}
\alias{IsPrime}
\title{Is a prime number?}
\usage{
IsPrime(n)
}
\arguments{
\item{n}{A positive integer.}
}
\value{
A boolean.
}
\description{
Is n a prime number?
}
\examples{
IsPrime(1)         # FALSE
IsPrime(5)         # TRUE
IsPrime(59999999)  # TRUE
}

